Remove Dockerfile, run locally with plain conda environments
The jupyter base image we were using hadn't been updated to Python 3.13
yet. There were file permission errors because the container ran as
a different user meaning modifications to notebooks couldn't be saved.
The environment contained many superfluous tools such as Julia. The
Docker environment didn't offer much improvement over a plain conda
environment and caused issues that were difficult to rectify.

Instead instructions have been added on how to run this directly using
conda. This provides none of the down sides of the Docker container and
allows us to run the latest version of everything. The instructions to
run Jupyter Lab are longer, but most of that is setting up a conda
environment. This is a process that many users will already be familiar
with.