emsarray 0.9.0 (#6)
* Update to emsarray 0.9.0, python 3.13

* Clear notebook outputs

* Tidy up the notebooks some

* Remove Dockerfile, run locally with plain conda environments

The jupyter base image we were using hadn't been updated to Python 3.13
yet. There were file permission errors because the container ran as
a different user meaning modifications to notebooks couldn't be saved.
The environment contained many superfluous tools such as Julia. The
Docker environment didn't offer much improvement over a plain conda
environment and caused issues that were difficult to rectify.

Instead instructions have been added on how to run this directly using
conda. This provides none of the down sides of the Docker container and
allows us to run the latest version of everything. The instructions to
run Jupyter Lab are longer, but most of that is setting up a conda
environment. This is a process that many users will already be familiar
with.

* Use Binder links for notebooks

The Nectar platform only supports Python 3.10 currently, but emsarray,
numpy, xarray etc have all dropped support for in accordance with
SPEC-0000

https://scientific-python.org/specs/spec-0000/

* Fix conda instructions

* Fix TOC links in README.md